File::CAS::File now supports external buffer modifications
Previously, making changes to the length of $file->buffer would
cause bugs related to position and EOF.  Now, the position is
relative to the end of the buffer, rather than the beginning,
so reads will proceed as expected even if the buffer changed size.

Also, changed Handle objects to hold a strong reference to the $file,
so that "$cas->get($hash)->newHandle" will work. In the process,
I renamed "handle" to "newhandle".