Implemented Dir::Minimal encoding
File::CAS::Dir::Minimal contains a listing of the bare minimum
metadata needed to recreate a file tree, minus any permissions
or ownership.  I created it mostly for testing purposes, but it
might prove to be efficient enough that you could use it for
small lightweight backups inbetween the big complete backups.

Also revised DirScan to generate some different hash keys,
and added a new base class for dir entries that implements
a bunch of dummy accessors that return undef.