Finished re-work on File/CAS.pm itself, and updated lots of docs
File::CAS constructor is now updated to work with the rest
of the package.  I have logic to splice out the parameters
intended for the store, and have it create a default scanner
object for the putDir operation.

The Scanner module (formerly DirScan) now has a storeDir
method which recursively stores files and directories.
This is the bulk of a backup operation.