Added new Dir::Entry method 'asHash'
Now, you can take any Dir::Entry and call 'asHash' to get a simple
hash and be able to query what keys are available.

Also, for efficiency, Dir::Entry is now a blessed ref to a hashref,
so that it can return the un-blessed hash.  Its accessors are now
implemented as ->asHash->{field} so that they can work correctly for
all subclesses.