File::CAS::Scanner now has scanDirEnt and getConfig
This allows us to get an entry for a single item in a directory, which
is needed when grabbing arbitrary directories from the filesystem to be
stored as subdirectories in the virtual hierarchy.

the getConfig method was added for compatibility with App::Casbak, in
order to get a hash of parameters needed to re-create a File::CAS object
and sub-objects.

I also made the uidCache and gidCache into attributes rather than starting
from scratch on each directory.  I did not yet add methods to clear them.