Added getConfig and changed handling of default path in Store::Simple
Store::Simple now determined its path by applying 'pathBase' if 'path' is
relative.  pathBase defaults to the current directory.  This was added so
that App::Casbak can use a relative path of "./store" while still changing
the 'pathBase' on each invocation, and so the parameter 'path' can be saved
into the configuration for the backup.

The new method pathReal can be used to get the actual absolute path of the
store's files.

I also added getConfig, which returns parameters needed to re-instantiate
the store.