Changed semantics of File::CAS constructor
File::CAS no longer merges its parameters with those of the store.
Store parameters can still be specified as a sub-hash named ->{store} with
a field "CLASS" to indicate which type of store should be constructed,
but this field must now be an absolute pathname rather than a suffix of
File::CAS::Store.  Likewise, parameters for the Scanner can be specified in
a sub-hash ->{scanner}.

I also implemented getOptions, needed for re-constructing the CAS object.