Removed the path canonicalization from Store::Simple
Previously, I had tried to always use a fully resolved path
within Store::Simple to protect it against changes to the
current directory.  However, I decided that the module should
not be in charge of this, and the burden should be on the
user to create absolute paths, if desired.

In part, this was because File::Spec->rel2abs fails the taint
check when I later create files based on it.