Gave App::Casbak a nice overhaul, and a new cleaner design
App::Casbak is not based on Moo, and uses a lazy-built "config"
attribute, which contains the constructor parameters for all the
other lazy-built fields.  This design allows you to create
instances of App::Casbak even without an actual backup directory
and still get all the functionality.

Next, the initialization routine became much easier and
straightforward, since it can create a test object using a
configuration, and if it works, save that configuration to be
used again as-is.

I wrote test cases for the canonical_date function, and fixed
a few bugs.