Finished the wave of changes caused by splitting DirCodec from Dir
Everything is now updated, and I extended the test cases significantly on
the DirCodec subclasses.

I actually substantially altered the encoding of DirCodec::Unix, because
it turned out to have bugs related to the default-value feature.  The new
encoding puts the least-used fields last, and shortens the array to omit
any trailing fields which are undef.

The as_hash method of each DirCodec's DirEntry objects now only includes
keys for fields with defined values.