Reworked Unicode handling AGAIN
Now, all strings are assumed to have the codepoints that the user intends
to be stored, and can be freely extracted as either Unicode or Latin1,
whichever the decoder thinks is appropriate.  *except* instances of
NonUnicode, which are considered to be bytes always and forever, and come
back out as NonUnicode instances.

The idea now is that *if* the person populating the filesystem wants to
try presenting names as unicode, they can use instances of NonUnicode to
handle the problematic filenames if any are encountered.