Reworked the internal implementation of path overrides
Nodes are now linked with weak-references, which become regular references
if the node is altered.  Path objects now resolve to an array of nodes, and
hold onto the node references.  When a path changes, all sub-path nodes are
flagged as invalid.  When path objects check their node array, they see that
the leaf node has become invalid, and re-resolve.

This change allows paths to be partially-resolved, so that calling
  while ($path->readdir) {
    print $path->path($_)->type;
  }
is O(n) instead of O(n*m) where m is the length of the path.